pyproject.toml
[project]
name = "pythonlearnproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
reamde = "README.md"
dependencies = [
    "aisuite>=0.1.6",
    "anthropic>=0.75.0",
    "docstring-parser>=0.17.0",
    "duckdb>=1.4.3",
    "fastapi>=0.128.0",
    "google-genai>=1.56.0",
    "ipywidgets>=8.1.8",
    "jinja2>=3.1.6",
    "jupyter-server>=2.17.0",
    "jupyterlab>=4.5.1",
    "langchain>=1.2.0",
    "langchain-google-genai>=4.1.2",
    "markdown>=3.10",
    "matplotlib>=3.10.8",
    "mistralai>=1.10.0",
    "nbclassic>=1.3.3",
    "notebook>=7.5.1",
    "numpy>=2.4.0",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "pandas-stubs~=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "qrcode>=8.2",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.45",
    "tabulate>=0.9.0",
    "tavily-python>=0.7.17",
    "tensorflow>=2.20.0",
    "textstat>=0.7.12",
    "tinydb>=4.8.2",
    "utils>=1.0.2",
    "uvicorn>=0.40.0",
    "vertexai>=1.71.1",
    "wikipedia>=1.4.0",
]
